refactor: insert peer relation code move to relation domain
Due to migration of peer relations and ensure the relation is created
with the correct relation id. The insert peer relations code is being
moved to a new method in the relation domain: CreatePeerRelations.
