+++ domain/relation/state/types.go
@@ -290,3 +290,12 @@ type watcherMapperData struct {
 }
 
 type uuids []string
+
+// peerEndpoint represents a structure for defining a peer application endpoint
+// with a UUID and a name.
+type peerEndpoint struct {
+	// UUID is the unique identifier of the peer endpoint.
+	UUID corerelation.EndpointUUID `db:"uuid"`
+	// Name is the human-readable name of the peer endpoint.
+	Name string `db:"name"`
+}
